requirements.txt
pandas==2.2.3
python-decouple==3.8
pytz==2025.2
Requests==2.32.3
streamlit==1.44.1
